+++ components/radarChart/radarChart.types.ts
@@ -25,10 +25,31 @@ interface IRadarChartDefaultProps extends IRadarCompoundProps {
   defaultData: DataPoint[]
 }
 
+interface IAxisMaps {
+  axis: string
+  value: number
+  order: number
+}
+
+interface OtherKeywordPercents {
+  [key: string]: number
+}
+
+interface IRadarChartContainerProps {
+  isRegistered: boolean
+  originKeywordPercents: OtherKeywordPercents
+  otherKeywordPercents: OtherKeywordPercents
+  frameSize: number
+  radarSize: number
+  framePadding: number
+}
+
 export type {
   DataPoint,
   IRadarChartProps,
   IRadarCompoundProps,
   IRadarChartDraggableProps,
   IRadarChartDefaultProps,
+  IAxisMaps,
+  IRadarChartContainerProps,
 }
