+++ components/radarChart/radarChartContainer.tsx
@@ -8,26 +8,11 @@ import useBreakpoint from '@/hooks/useBreakpoint.hooks'
 
 import { rotateOffsetRatioMap } from './helpers/axis.helpers'
 import { RadarChart } from './radarChart'
-import { DataPoint } from './radarChart.types'
-
-interface IAxisMaps {
-  axis: string
-  value: number
-  order: number
-}
-
-interface OtherKeywordPercents {
-  [key: string]: number
-}
-
-interface IRadarChartContainerProps {
-  isRegistered: boolean
-  originKeywordPercents: OtherKeywordPercents
-  otherKeywordPercents: OtherKeywordPercents
-  frameSize: number
-  radarSize: number
-  framePadding: number
-}
+import {
+  DataPoint,
+  IAxisMaps,
+  IRadarChartContainerProps,
+} from './radarChart.types'
 
 export default function RadarChartContainer({
   isRegistered,
